+++ src/main/xml/XMLDeserializer.java
@@ -76,7 +76,7 @@ public class XMLDeserializer {
 		double latitude = Double.parseDouble(element.getAttribute("latitude"));
 		double longitude = Double.parseDouble(element.getAttribute("longitude"));
 	
-		return new Intersection((int) id, latitude, longitude);
+		return new Intersection( id, latitude, longitude);
 	}
 	
 
@@ -90,8 +90,8 @@ public class XMLDeserializer {
     private static Section createSection(Element elt, Plan plan) throws XMLException{ // TODO : naming
     	long idDeparture = Long.parseLong(elt.getAttribute("origine"));
     	long idArrival = Long.parseLong(elt.getAttribute("destination"));
-   		Intersection departure = plan.getIntersectionById((int) idDeparture); // XXX : is this OK ?
-   		Intersection arrival = plan.getIntersectionById((int) idArrival);
+   		Intersection departure = plan.getIntersectionById( idDeparture); // XXX : is this OK ?
+   		Intersection arrival = plan.getIntersectionById( idArrival);
    		
    		String streetName = elt.getAttribute("nomRue");
    		
@@ -136,7 +136,7 @@ public class XMLDeserializer {
 	}
 	
 	private static Delivery createDelivery(Element elt, Plan plan) {
-		Intersection departure = plan.getIntersectionById((int) Long.parseLong(elt.getAttribute("adresse")));;
+		Intersection departure = plan.getIntersectionById( Long.parseLong(elt.getAttribute("adresse")));;
 		
 		if (elt.hasAttribute("duree")) { // Actual delivery
 			int duration = Integer.parseInt(elt.getAttribute("duree"));
